Support Trilio encryption options
Add suppot for Trilio encryption options TRILIO_ENCRYPTION_SUPPORT
and OPENSTACK_ENCRYPTION_SUPPORT. TRILIO_ENCRYPTION_SUPPORT should
set to True for trilio > 4.2. OPENSTACK_ENCRYPTION_SUPPORT should
be set to True if Barbican is present in the deployment. It would
be nice if Barbicans presence was detected via a relation but since
this option is going in Trilio 5.X exposing this as a charm
option is a tactical fix. As part of the work the charm classes
were restructured a little as this charm does not support 4.0.

func-test-pr: https://github.com/openstack-charmers/zaza-openstack-tests/pull/856

Change-Id: I29e83cab68ba670204f27f15010459836d8f8160